pyproject.toml
[tool.poetry]
name = "font-search-api"
version = "0.1.0"
description = ""
authors = ["masashi kishimoto <drehbleistift@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.3"
fastapi = "^0.75.1"
albumentations = "^1.1.0"
Pillow = "^9.1.0"
mangum = "^0.14.1"
uvicorn = "^0.17.6"
google-cloud-vision = "^2.7.2"
boto3 = "^1.23.0"
python-dotenv = "^0.20.0"
azure-cognitiveservices-vision-computervision = "^0.9.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.2"
torch = "^1.11.0"
torchvision = "^0.12.0"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
